Added -Wextra and -Wshorten-64-to-32 and fixed resulting errors. (#289)
* Added -Wextra and -Wshorten-64-to-32 and fixed resulting errors.

* Disable -Wshorten-32-to-64 since Kokoro is missing Clang.

* Fixed -Wextra warnings for gcc.

* Reordered UPB_UNUSED() to come after declarations.

* Added another -pedantic fix and log CC version.

* Fix compile error and conditionally run use_bazel.sh.

* Moved set -e after use_bazel.sh.

* Fixed typo in conditional.

+++ tests/conformance_upb.c
@@ -138,7 +138,7 @@ void serialize_json(const upb_msg *msg, const upb_msgdef *m, const ctx *c) {
   upb_status_clear(&status);
   len = upb_json_encode(msg, m, c->symtab, opts, NULL, 0, &status);
 
-  if (len == -1) {
+  if (len == (size_t)-1) {
     const char *inerr = upb_status_errmsg(&status);
     size_t len = strlen(inerr);
     char *err = upb_arena_malloc(c->arena, len + 1);
